refactor(blorktools): further splitting of rig panel
